Working success😊: Add TCP socket initialization and connection methods to ReactNativeTcpWin module

+++ windows/ReactNativeTcpWin/ReactNativeTcpWin.h
@@ -28,16 +28,16 @@ struct ReactNativeTcpWin
   double add(double a, double b) noexcept;
 
   REACT_SYNC_METHOD(InitializeTcpSocket)
-  int InitializeTcpSocket () noexcept;
+  double InitializeTcpSocket () noexcept;
 
   REACT_SYNC_METHOD(ConnectToServer)
-  int ConnectToServer(std::string ipAddress, std::string port) noexcept;
+  double ConnectToServer(std::string ipAddress, std::string port) noexcept;
 
   REACT_SYNC_METHOD(SendMessage)
-  int SendMessage(std::string message) noexcept;
+  double SendMessage(std::string message) noexcept;
 
   REACT_SYNC_METHOD(DisconnectFromServer)
-  int DisconnectFromServer() noexcept;
+  double DisconnectFromServer() noexcept;
 
 private:
   React::ReactContext m_context;
